Fixing the errors that appears
