+++ src/Routes/Calendar/ScheduleForm/Containers/DataCollection/OptionsForm.js
@@ -6,7 +6,7 @@ const HourValue = (props) => {
     const Hours = ['01','02','03','04','05','06','07','08','09','10','11','12']
     const DailyHours = (
         [Hours.map((Hour, index) => 
-            <option disabled={props.DisabledHour}  key={index.id+1} value={Hour}>{Hour}</option>)]
+            <option disabled={props.DisabledHour}  key={index.id + 1} value={Hour}>{Hour}</option>)]
         )
     return(DailyHours)
 }
@@ -16,7 +16,7 @@ const MinutsValue = (props) => {
     const Minuts = ['05','10','15','20','25','30','35','40','45','50','55']
     const DailyMinut =(
         [Minuts.map((Minut, index) => 
-            <option disabled={props.DisabledMinuts} key={index.id+1} value={Minut}>{Minut}</option>)]
+            <option disabled={props.DisabledMinuts} key={index.id + 1} value={Minut}>{Minut}</option>)]
         )
     return(DailyMinut)
 }
@@ -26,7 +26,7 @@ const TwelveHourSystemValue = (props) => {
     const TwelveHourSystem = ['AM','PM']
     const DayDivided =(
         [TwelveHourSystem.map((DayTime, index) => 
-            <option disabled={props.DisabledHourSystem} key={index.id+1} value={DayTime}>{DayTime}</option>)]
+            <option disabled={props.DisabledHourSystem} key={index.id + 1} value={DayTime}>{DayTime}</option>)]
         )
     return(DayDivided)
 }
@@ -36,7 +36,7 @@ const NumberOfYearsValue = (props) => {
                             '2027', '2028', '2029', '2030', '2031', '2032','2033', 
                             '2034','2035', '2036', '2037', '2038', '2039','2040']
     const YearContainer = (
-        [NumberOfYears.map((year, index) => <option  disabled={props.DisabledYear}  key={index.id+1} value={year}>{year}</option>)]
+        [NumberOfYears.map((year, index) => <option  disabled={props.DisabledYear}  key={index.id + 1} value={year}>{year}</option>)]
     )
     
     return(
@@ -47,7 +47,7 @@ const NumberOfMonthsValue = (props) => {
     //this is for the months in a year 
     const NumberOfMonths = ['01','02','03','04','05','06','07','08','09','10','11','12']
     const MonthContainer = (
-        [NumberOfMonths.map((month, index) => <option disabled={props.DisabledMonth} key={index.id+1}  value={month}>{month}</option>)]
+        [NumberOfMonths.map((month, index) => <option disabled={props.DisabledMonth} key={index.id + 1}  value={month}>{month}</option>)]
         )
     return(MonthContainer)
 }
@@ -57,7 +57,7 @@ const NumberOfDaysValue = (props) => {
                         '11','12','13','14','15','16','17','18','19','20',
                         '21','22','23','24','25','26','27','28','29','30','31']
     const DayContainer = (
-        [NumberOfDays.map((days,index ) => <option disabled={props.DisabledDay} key={index.id+1} value={days}>{days}</option>)]
+        [NumberOfDays.map((days,index ) => <option disabled={props.DisabledDay} key={index.id + 1} value={days}>{days}</option>)]
         )
     return(DayContainer)
 }
@@ -67,7 +67,7 @@ const ColorOptionsValue = () => {
     const ColorChoose = ['Choose Color','BlueViolet','Black','CornflowerBlue','DarkMagenta','DimGrey',
                             'DodgerBlue','Fuchsia','MidnightBlue','Purple','Salmon']
     const ColorContainer = (
-        [ColorChoose.map((Colors,index ) => <option  key={index.id+1} value={Colors}>{Colors}</option>)]
+        [ColorChoose.map((Colors,index ) => <option  key={index.id + 1} value={Colors}>{Colors}</option>)]
         )
     return(ColorContainer)
 }
